server/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn==0.30.6

# HTTP
httpx==0.27.2
python-multipart==0.0.9

# AI / ML
google-generativeai==0.8.3
torch==2.4.1
torchvision==0.19.1
transformers==4.45.1
ultralytics==8.2.103
mediapipe==0.10.14
librosa==0.10.2
opencv-python==4.10.0.84
Pillow==10.4.0

# Video Processing
yt-dlp==2024.9.27

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
redis==5.1.1

# Utils
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
